pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
pydantic-settings = "^2.6.1"
uvicorn = "^0.32.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
uuid = "^1.30"
python-jose = "^3.3.0"
passlib = "^1.7.4"
celery = "^5.4.0"
aio-pika = "^9.5.0"
redis = "^5.2.0"
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[tool.pytest.ini_options]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
